Implemented support for GitLab cloud. Updated gitlab4j version. (#91)
* Implemented support for GitLab cloud. Updated gitlab4j version.

* Fixed long value in merge request handler

+++ src/main/java/com/microfocus/octane/gitlab/helpers/TestResultsHelper.java
@@ -26,7 +26,7 @@ public class TestResultsHelper {
         return Paths.get(rootFolder, testResultFolder).toFile();
     }
 
-    public static File getMQMTestResultsFilePath(Integer projectId, Integer jobId,String rootFolderPath) {
+    public static File getMQMTestResultsFilePath(long projectId, long jobId,String rootFolderPath) {
 
         if(rootFolderPath == null ){
             rootFolderPath ="";
